requirements.txt
streamlit==1.36.0
numpy==1.26.4
tensorflow==2.15.0
opencv-python-headless
pillow==10.4.0